Moved RestClient to Util
