+++ src/main/java/me/philippheuer/twitch4j/TwitchClient.java
@@ -10,8 +10,8 @@ import me.philippheuer.twitch4j.chat.IrcClient;
 import me.philippheuer.twitch4j.chat.commands.CommandHandler;
 import me.philippheuer.twitch4j.endpoints.*;
 import me.philippheuer.twitch4j.events.EventDispatcher;
-import me.philippheuer.twitch4j.helper.HeaderRequestInterceptor;
-import me.philippheuer.twitch4j.helper.RestClient;
+import me.philippheuer.util.rest.HeaderRequestInterceptor;
+import me.philippheuer.util.rest.RestClient;
 import me.philippheuer.twitch4j.pubsub.TwitchPubSub;
 import me.philippheuer.twitch4j.streamlabs.StreamlabsClient;
 import org.springframework.util.Assert;
