+++ src/main/java/me/philippheuer/twitch4j/auth/OAuthStreamlabs.java
@@ -5,7 +5,7 @@ import lombok.Setter;
 import me.philippheuer.twitch4j.auth.model.OAuthCredential;
 import me.philippheuer.twitch4j.auth.model.OAuthRequest;
 import me.philippheuer.twitch4j.auth.model.streamlabs.Authorize;
-import me.philippheuer.twitch4j.helper.WebsiteUtils;
+import me.philippheuer.util.desktop.WebsiteUtils;
 import me.philippheuer.twitch4j.streamlabs.enums.StreamlabsScopes;
 import me.philippheuer.twitch4j.streamlabs.model.User;
 import org.springframework.util.LinkedMultiValueMap;
@@ -14,7 +14,6 @@ import org.springframework.web.client.RestTemplate;
 
 import java.util.Arrays;
 import java.util.Optional;
-import java.util.UUID;
 import java.util.stream.Collectors;
 
 @Getter
