+++ src/main/java/me/philippheuer/twitch4j/endpoints/KrakenEndpoint.java
@@ -5,22 +5,9 @@ import lombok.Getter;
 import lombok.Setter;
 import me.philippheuer.twitch4j.TwitchClient;
 import me.philippheuer.twitch4j.auth.model.OAuthCredential;
-import me.philippheuer.twitch4j.enums.TwitchScopes;
-import me.philippheuer.twitch4j.events.Event;
-import me.philippheuer.twitch4j.events.event.DonationEvent;
-import me.philippheuer.twitch4j.events.event.FollowEvent;
-import me.philippheuer.twitch4j.exceptions.ChannelCredentialMissingException;
-import me.philippheuer.twitch4j.exceptions.ChannelDoesNotExistException;
 import me.philippheuer.twitch4j.exceptions.RestException;
-import me.philippheuer.twitch4j.helper.HeaderRequestInterceptor;
-import me.philippheuer.twitch4j.helper.QueryRequestInterceptor;
 import me.philippheuer.twitch4j.model.*;
-import me.philippheuer.twitch4j.streamlabs.endpoints.DonationEndpoint;
-import me.philippheuer.twitch4j.streamlabs.model.Donation;
-import org.springframework.util.Assert;
 import org.springframework.web.client.RestTemplate;
-
-import java.util.*;
 
 @Getter
 @Setter
