--- src/main/java/me/philippheuer/twitch4j/helper/package-info.java
@@ -1,4 +1,0 @@
-/**
- * This package contains all helper classes, that don't belong anywhere.
- */
-package me.philippheuer.twitch4j.helper;
