+++ src/main/java/me/philippheuer/twitch4j/streamlabs/StreamlabsClient.java
@@ -5,8 +5,8 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.Singular;
 import me.philippheuer.twitch4j.auth.model.OAuthCredential;
-import me.philippheuer.twitch4j.helper.HeaderRequestInterceptor;
-import me.philippheuer.twitch4j.helper.RestClient;
+import me.philippheuer.util.rest.HeaderRequestInterceptor;
+import me.philippheuer.util.rest.RestClient;
 import me.philippheuer.twitch4j.streamlabs.endpoints.AlertEndpoint;
 import me.philippheuer.twitch4j.streamlabs.endpoints.DonationEndpoint;
 import me.philippheuer.twitch4j.streamlabs.endpoints.UserEndpoint;
