+++ src/main/java/me/philippheuer/twitch4j/streamlabs/endpoints/AlertEndpoint.java
@@ -5,8 +5,7 @@ import lombok.Getter;
 import lombok.Setter;
 import me.philippheuer.twitch4j.auth.model.OAuthCredential;
 import me.philippheuer.twitch4j.exceptions.RestException;
-import me.philippheuer.twitch4j.helper.LoggingRequestInterceptor;
-import me.philippheuer.twitch4j.helper.QueryRequestInterceptor;
+import me.philippheuer.util.rest.LoggingRequestInterceptor;
 import me.philippheuer.twitch4j.streamlabs.StreamlabsClient;
 import me.philippheuer.twitch4j.streamlabs.model.AlertCreate;
 import org.springframework.util.LinkedMultiValueMap;
