+++ src/main/java/me/philippheuer/twitch4j/streamlabs/endpoints/DonationEndpoint.java
@@ -6,8 +6,8 @@ import lombok.Setter;
 import me.philippheuer.twitch4j.auth.model.OAuthCredential;
 import me.philippheuer.twitch4j.exceptions.CurrencyNotSupportedException;
 import me.philippheuer.twitch4j.exceptions.RestException;
-import me.philippheuer.twitch4j.helper.LoggingRequestInterceptor;
-import me.philippheuer.twitch4j.helper.QueryRequestInterceptor;
+import me.philippheuer.util.rest.LoggingRequestInterceptor;
+import me.philippheuer.util.rest.QueryRequestInterceptor;
 import me.philippheuer.twitch4j.streamlabs.StreamlabsClient;
 import me.philippheuer.twitch4j.streamlabs.model.Donation;
 import me.philippheuer.twitch4j.streamlabs.model.DonationCreate;
