+++ src/main/java/me/philippheuer/util/desktop/WebsiteUtils.java
@@ -1,4 +1,4 @@
-package me.philippheuer.twitch4j.helper;
+package me.philippheuer.util.desktop;
 
 import java.awt.*;
 import java.net.URI;
