+++ src/main/java/me/philippheuer/util/desktop/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package contains desktop utilities.
+ */
+package me.philippheuer.util.desktop;
