+++ src/main/java/me/philippheuer/util/rest/RestErrorHandler.java
@@ -1,4 +1,4 @@
-package me.philippheuer.twitch4j.helper;
+package me.philippheuer.util.rest;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import me.philippheuer.twitch4j.exceptions.RestException;
