+++ src/main/java/me/philippheuer/util/rest/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package contains the rest client, request interceptors and related classes.
+ */
+package me.philippheuer.util.rest;
