+++ src/main/java/org/springframework/social/support/package-info.java
@@ -1,4 +1,4 @@
 /**
- * This package contains a few classes of the spring social support library.
+ * This package contains a few classes of the rest social support library.
  */
 package org.springframework.social.support;
